+++ frontend/src/app/(routes)/staking/validator/[validator]/components/TableHeader.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const TableHeader = ({ title }: { title: string }) => {
+  return (
+    <th>
+      <div className="min-h-[17px] flex items-center text-sm not-italic font-normal leading-[normal]">
+        {title}
+      </div>
+    </th>
+  );
+};
+
+export default TableHeader;
